requirements.txt
dataclasses
jwcrypto
pyjwt
requests
